@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "description": "Quick start Rate Provider App",
   "main": "index.js",
+  "private": true,
   "license": "Apache-2.0",
   "dependencies": {
     "@plexus-interop/client": "0.1.0",
@@ -20,7 +21,7 @@
   "devDependencies": {
     "browserify": "^14.4.0",
     "copyfiles": "^1.2.0",
-    "electron": "1.7.12",
+    "electron": "1.8.4",
     "trash": "^4.1.0",
     "trash-cli": "^1.4.0",
     "typescript": "2.6.1"
