@@ -5,6 +5,7 @@
   "main": "dist/main/src/index.js",
   "description": "Integration tests for Plexus Web Interop",
   "license": "Apache-2.0",
+  "private": true,
   "scripts": {
     "prebuild": "trash dist",
     "build": "yarn build:main && npm run copy-assets && npm run build-host-proxy",
@@ -50,7 +51,7 @@
     "chokidar": "^1.7.0",
     "copyfiles": "^1.2.0",
     "core-js": "^2.5.3",
-    "electron": "1.7.12",
+    "electron": "1.8.4",
     "electron-mocha": "^4.0.2",
     "glob": "^7.1.2",
     "http-server": "^0.11.1",
@@ -80,9 +81,6 @@
   "jest": {
     "testEnvironment": "node",
     "setupTestFrameworkScriptFile": "<rootDir>/setup-jasmine.js",
-    "transform": {
-      ".(ts|tsx)": "<rootDir>/preprocessor.js"
-    },
     "moduleFileExtensions": [
       "ts",
       "js"
