@@ -1,4 +1,4 @@
-/*
+/**
  * Copyright 2017 Plexus Interop Deutsche Bank AG
  * SPDX-License-Identifier: Apache-2.0
  *
@@ -14,19 +14,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-const tsc = require('typescript');
-const tsConfig = require('./tsconfig.json');
+import { BaseJavaGenCommand } from './BaseJavaGenCommand';
+import { baseDir, out, plexusEntryPoint } from './DefaultOptions';
+import { Option } from './Option';
 
-module.exports = {
-    process(src, path) {
-        if (path.endsWith('.ts') || path.endsWith('.tsx')) {
-            return tsc.transpile(
-                src,
-                tsConfig.compilerOptions,
-                path,
-                []
-            );
-        }
-        return src;
-    },
-};
+export class GenProtoCommand extends BaseJavaGenCommand {
+    
+    public plexusGenArgs: (opts: any) => string[] = opts => {
+        return ['--type=proto', ...this.optionArgs(opts)];
+    }
+
+    public name = () => 'gen-proto';
+
+    public options: () => Option[] = () => [baseDir(), out(), plexusEntryPoint()];
+
+}
