@@ -1,4 +1,4 @@
-/*
+/**
  * Copyright 2017 Plexus Interop Deutsche Bank AG
  * SPDX-License-Identifier: Apache-2.0
  *
@@ -14,19 +14,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-const tsc = require('typescript');
-const tsConfig = require('./tsconfig.json');
+export interface Option {
+    shortName: string;
+    longName: string;
+    exampleValue: string;
+    description?: string;
+    isRequired?: boolean;
+    defaultValue?: any;
+}
 
-module.exports = {
-    process(src, path) {
-        if (path.endsWith('.ts') || path.endsWith('.tsx')) {
-            return tsc.transpile(
-                src,
-                tsConfig.compilerOptions,
-                path,
-                []
-            );
-        }
-        return src;
-    },
-};
+export function getFlags(o: Option): string {
+    return !!o.isRequired ? 
+        `-${o.shortName}, --${o.longName} <${o.longName}>`
+            : `-${o.shortName}, --${o.longName}`;
+}
