@@ -18,7 +18,6 @@ import { GenericClientApi } from "./GenericClientApi";
 import { GenericClient } from "../../../client/generic/GenericClient";
 import { ServiceDiscoveryRequest } from "@plexus-interop/client-api";
 import { ServiceDiscoveryResponse } from "@plexus-interop/client-api";
-import { Observer } from "@plexus-interop/common";
 import { ClientDtoUtils } from "./../../ClientDtoUtils";
 import { StreamingInvocationClient } from "./../streaming/StreamingInvocationClient";
 import { StreamingInvocationClientImpl } from "./../streaming/StreamingInvocationClientImpl";
@@ -29,7 +28,7 @@ import { InvocationRequestInfo } from "@plexus-interop/protocol";
 import { Logger, LoggerFactory, Arrays } from "@plexus-interop/common";
 import { MarshallerProvider } from "../io/MarshallerProvider";
 import { Completion } from "@plexus-interop/client-api";
-import { DelegateChannelObserver, UniqueId } from "@plexus-interop/transport-common";
+import { UniqueId } from "@plexus-interop/transport-common";
 import { ProvidedMethodReference } from "@plexus-interop/client-api";
 import { Invocation } from "../../generic/Invocation";
 import { MethodDiscoveryRequest } from "@plexus-interop/client-api";
