@@ -16,7 +16,7 @@
  */
 import { Invocation } from "../../client/generic/Invocation";
 import { GenericInvocation } from "./GenericInvocation";
-import { UniqueId, ChannelObserver } from "@plexus-interop/transport-common";
+import { UniqueId } from "@plexus-interop/transport-common";
 import { AnonymousSubscription } from "../../client/api/AnonymousSubscription";
 import { clientProtocol as plexus, InvocationMetaInfo } from "@plexus-interop/protocol";
 import { ProvidedMethodReference } from "@plexus-interop/client-api";
