@@ -18,10 +18,10 @@ import { GenericInvocation } from "../../src/client/generic/GenericInvocation";
 import { InvocationState } from "../../src/client/generic/InvocationState";
 import { clientProtocol as plexus } from "@plexus-interop/protocol";
 import { InvocationMetaInfo } from "@plexus-interop/protocol";
-import { UniqueId, DelegateChannelObserver } from "@plexus-interop/transport-common";
+import { UniqueId } from "@plexus-interop/transport-common";
 import { ClientProtocolHelper as modelHelper } from "@plexus-interop/protocol";
 import { randomPayload } from "../utils";
-import { AsyncHelper, Observer } from "@plexus-interop/common";
+import { AsyncHelper } from "@plexus-interop/common";
 import { CancellationToken } from "@plexus-interop/common";
 import { createInvocationInfo } from "./client-mocks";
 import { BufferedChannel } from "./client-mocks";
